+++ src/main/java/com/zipcodewilmington/scientificcalculator/MainApplication.java
@@ -30,7 +30,6 @@ public class MainApplication {
         CoreFeatures x = new CoreFeatures();
       //  x.calc();
 
-        x.setDisMode();
 
 
 //        ScientificFeatures x = new ScientificFeatures();
